+++ universum_log_collapser/plugin/src/main/java/io/jenkins/plugins/universum_log_collapser/AnnotatorFactory.java
@@ -8,6 +8,6 @@ import hudson.console.ConsoleAnnotatorFactory;
 public class AnnotatorFactory<T> extends ConsoleAnnotatorFactory<T> {
     @Override
     public Annotator newInstance(T context) {
-        return new Annotator(context);
+        return new Annotator();
     }
 }