+++ CHANGELOG.rst
@@ -1,6 +1,27 @@
 Change log
 ==========
 
+0.18.3 (2020-01-10)
+-------------------
+
+New features
+~~~~~~~~~~~~
+
+* **launcher command line arguments renamed**
+   * see :doc:`'Output' and 'Configuration execution' arguments <args>` for new options list
+   * old options `--launcher-config-path` and `--launcher-output` are still supported
+     but not recommended to use any more
+* **launcher:**  add '--filter' option filtering steps to be executed
+* **nonci:** add :doc:`nonci <args_nonci>` mode of running Universum
+* **jenkins_plugin:** expand failed steps by default
+* **test:** add Java & JS tests for Jenkins plugin
+
+Bug fixes
+~~~~~~~~~
+
+* **jenkins_plugin:** fix collapsing with timestamps plugin usage
+
+
 0.18.2 (2019-10-09)
 -------------------